dog struct

+++ 0x0E-structures_typedef/dog.h
@@ -7,12 +7,12 @@
  * @age: Age of dog
  * @owner: Name of owner
  *
- * Description: A structure for a dog
+ * Description: Data structure for a dog
  */
 struct dog
 {
   char *name;
-  int age;
+  float age;
   char *owner;
 };
